[FSI-552] Use the api server CA to manage certificates (#8)
This is a proof of concept of certificate management for mutating webhook.

## Objectives 🎯

The objective of this PR is to have certificate management that is secure and allows the service to be scaled to as many replicas as needed. Bonus features to have: transparency from the user's point of view in the setup process.

It should also allow users to be able to provide their own key/pair & teach them how to patch the webhook registration with their own CA bundle.

## Automagic certificate management

### How it works? 🤔

The api server is used as a CA (certificate authority) to sign a certificate for the mutating webhook server. The signed certificate & key pair are put in a secret. This secret is then mounted in the pods and the webhook server loads them while starting.

This approach avoids having to run the webhook server in a highly privileged container and also to generate everything in the container that needs to identify itself 🌀.

The webhook server also listens for changes in the signed certificate & key pair to update its own configuration if they ever change. 

While the reloading is happening the webhook server should not receive requests, otherwise, it could have the wrong key pair loaded to serve it.

#### Certificate signing process 🖊📃

To handle the whole process, a Kubernetes **Job** is used. It's fired only once and works as follows:

1. A server key is generated
2. If there is any previous CSR (certificate signing request) for this key, it is deleted
3. A new CSR is generated for such key
4. The signature of the key is then approved
5. The server's certificate is fetched from the CSR and then encoded
6. A secret of type `tls` is created with the server certificate and key
7. The api server's CA bundle is fetched
8. The admission webhook registration for the webhook server is patched with the api server's CA bundle from the previous step

For the moment, this job is using `kubectl` directly, from a custom built Alpine image that also includes `openssl`. But if decided we could check if simple `curl`s could work.

## Manual certificate management

### How it works? 🤔

After the client already has the signed certificate/key pair:

1. The TLS secret should be created;
2. And the CA bundle should be updated in the webhook registration.

Proper instructions for this scenario should be present in the README of the project.

## Future improvements 🚧

1. If the api server's CA expires, we need to patch the admission webhook registration again with the new CA bundle
2. Add option of integrating with [cert-manager](https://github.com/jetstack/cert-manager)
3. Render the manifests with Helm?